+++ python/lsst/ts/wep/task/donutStamp.py
@@ -27,12 +27,16 @@ from typing import Optional
 import lsst.afw.image as afwImage
 import lsst.afw.table as afwTable
 import lsst.geom
-import lsst.obs.lsst as obs_lsst
 import numpy as np
 from lsst.afw.cameraGeom import FIELD_ANGLE, PIXELS
 from lsst.meas.algorithms.stamps import AbstractStamp
 from lsst.ts.wep.cwfs.compensableImage import CompensableImage
-from lsst.ts.wep.utils import DefocalType, FilterType, getFilterTypeFromBandLabel, getCameraFromButlerName
+from lsst.ts.wep.utils import (
+    DefocalType,
+    FilterType,
+    getCameraFromButlerName,
+    getFilterTypeFromBandLabel,
+)
 
 
 @dataclass